feat: enable "open with" functionality in PWA
When Mainsail is installed as an app on a desktop, it should get configured as an application that supports opening ".gcode" files. So you can either directly association gcode files to it, or right-click > Open With > Mainsail. When you do this, the file will get uploaded to the server and then open the print start dialog so that you can review and start the print.